Refactor goal runtime state machine
