+++ src/goal-runtime-event-handler-types.ts
@@ -91,6 +91,49 @@ export interface GoalRuntimeEventHandlerDeps {
   resetErrorRecovery: () => void;
 }
 
+export interface StaleQueuedWorkRuntimePort {
+  clearActiveAccounting: () => void;
+  status: GoalRuntimeStatusPort;
+}
+
+export interface InputContextHandlerDeps extends StaleQueuedWorkRuntimePort {
+  runtimeState: GoalRuntimeState;
+  stateController: GoalStateController;
+  continuation: GoalRuntimeContinuationPort;
+  recoveryRuntime: RecoveryRuntimePort;
+  resetErrorRecovery: () => void;
+}
+
+export interface TurnHandlerDeps extends StaleQueuedWorkRuntimePort {
+  runtimeState: GoalRuntimeState;
+  stateController: GoalStateController;
+  continuation: GoalRuntimeContinuationPort;
+  goalAccounting: GoalAccountingPort;
+  recoveryRuntime: RecoveryRuntimePort;
+}
+
+export interface AgentHandlerDeps extends StaleQueuedWorkRuntimePort, RecoveryEventDeps {
+  runtimeState: GoalRuntimeState;
+  continuation: GoalRuntimeContinuationPort;
+  goalAccounting: GoalAccountingPort;
+  resetErrorRecovery: () => void;
+}
+
+export interface SessionHandlerDeps extends StaleQueuedWorkRuntimePort {
+  pi: ExtensionAPI;
+  runtimeState: GoalRuntimeState;
+  stateController: GoalStateController;
+  continuation: GoalRuntimeContinuationPort;
+  goalAccounting: GoalAccountingPort;
+  recoveryRuntime: RecoveryRuntimePort;
+  resetErrorRecovery: () => void;
+}
+
+export interface RecoveryEventDeps {
+  stateController: GoalStateController;
+  recoveryRuntime: RecoveryRuntimePort;
+}
+
 export type QueuedGoalWorkMessage = {
   role: string;
   customType?: string;
