+++ src/goal-runtime-event-handlers.ts
@@ -4,8 +4,12 @@ import { createSessionEventHandlers } from "./goal-runtime-session-handlers.js";
 import { createTurnEventHandlers } from "./goal-runtime-turn-handlers.js";
 import { createQueuedGoalWorkMessageIdResolver } from "./goal-runtime-event-utils.js";
 import type {
+  AgentHandlerDeps,
   GoalRuntimeEventHandlerDeps,
   GoalRuntimeEventHandlers,
+  InputContextHandlerDeps,
+  SessionHandlerDeps,
+  TurnHandlerDeps,
 } from "./goal-runtime-event-handler-types.js";
 
 export type {
@@ -22,10 +26,50 @@ export function createGoalRuntimeEventHandlers(
     deps.continuation,
   );
 
+  const staleQueuedWorkRuntime = {
+    clearActiveAccounting: deps.clearActiveAccounting,
+    status: deps.status,
+  };
+  const inputContextDeps: InputContextHandlerDeps = {
+    ...staleQueuedWorkRuntime,
+    runtimeState: deps.runtimeState,
+    stateController: deps.stateController,
+    continuation: deps.continuation,
+    recoveryRuntime: deps.recoveryRuntime,
+    resetErrorRecovery: deps.resetErrorRecovery,
+  };
+  const turnDeps: TurnHandlerDeps = {
+    ...staleQueuedWorkRuntime,
+    runtimeState: deps.runtimeState,
+    stateController: deps.stateController,
+    continuation: deps.continuation,
+    goalAccounting: deps.goalAccounting,
+    recoveryRuntime: deps.recoveryRuntime,
+  };
+  const agentDeps: AgentHandlerDeps = {
+    ...staleQueuedWorkRuntime,
+    runtimeState: deps.runtimeState,
+    stateController: deps.stateController,
+    continuation: deps.continuation,
+    goalAccounting: deps.goalAccounting,
+    recoveryRuntime: deps.recoveryRuntime,
+    resetErrorRecovery: deps.resetErrorRecovery,
+  };
+  const sessionDeps: SessionHandlerDeps = {
+    ...staleQueuedWorkRuntime,
+    pi: deps.pi,
+    runtimeState: deps.runtimeState,
+    stateController: deps.stateController,
+    continuation: deps.continuation,
+    goalAccounting: deps.goalAccounting,
+    recoveryRuntime: deps.recoveryRuntime,
+    resetErrorRecovery: deps.resetErrorRecovery,
+  };
+
   return {
-    ...createInputContextEventHandlers(deps, queuedGoalWorkMessageIdForRuntime),
-    ...createTurnEventHandlers(deps),
-    ...createAgentEventHandlers(deps),
-    ...createSessionEventHandlers(deps),
+    ...createInputContextEventHandlers(inputContextDeps, queuedGoalWorkMessageIdForRuntime),
+    ...createTurnEventHandlers(turnDeps),
+    ...createAgentEventHandlers(agentDeps),
+    ...createSessionEventHandlers(sessionDeps),
   };
 }
