+++ src/goal-runtime-input-context-handlers.ts
@@ -12,13 +12,13 @@ import { isCommandResumeQueuedGoalMessage } from "./queued-goal-messages.js";
 import { applyStaleQueuedWorkEffects } from "./goal-runtime-event-utils.js";
 import type {
   ContextEventResult,
-  GoalRuntimeEventHandlerDeps,
+  InputContextHandlerDeps,
   MessageStartEvent,
   QueuedGoalWorkMessageIdResolver,
 } from "./goal-runtime-event-handler-types.js";
 
 export function createInputContextEventHandlers(
-  deps: GoalRuntimeEventHandlerDeps,
+  deps: InputContextHandlerDeps,
   queuedGoalWorkMessageIdForRuntime: QueuedGoalWorkMessageIdResolver,
 ) {
   const { runtimeState, stateController, continuation, recoveryRuntime, status, resetErrorRecovery } = deps;
